+++ src/controllers/partners.py
@@ -1,22 +1,12 @@
-################################################################################
-## IMPORTS #####################################################################
-################################################################################
-
-
-import cgi
 import jinja2
 import os
-import webapp2
 
-from datetime import date, datetime, timedelta
-from google.appengine.api import images, users
+from google.appengine.api import users
 from google.appengine.ext import ndb
-from json import dumps
 from webapp2_extras.appengine.users import login_required
 
 from handler import CustomHandler
 from models import Student, Invitation, Partnership, Evaluation, Setting
-from src.helpers.helpers import query_to_dict, split_last
 from src.models.assignment import AssignmentModel
 from src.models.eval import EvalModel
 from src.models.invitation import InvitationModel
@@ -26,77 +16,10 @@ from src.models.student import StudentModel
 from src.send_mail import SendMail 
 
 
-################################################################################
-################################################################################
-################################################################################
-
-
-################################################################################
-## LOAD JINJA ##################################################################
-################################################################################
-
-
 JINJA_ENV = jinja2.Environment(
     loader = jinja2.FileSystemLoader(os.path.dirname('app.yaml')),
     extensions = ['jinja2.ext.autoescape'],
     autoescape=True)
-
-
-################################################################################
-################################################################################
-################################################################################
-
-
-################################################################################
-## HANDLERS ####################################################################
-################################################################################
-
-
-class BrowseForPartners(CustomHandler):
-
-    @login_required
-    def get(self):
-        quarter  = SettingModel.quarter()
-        year     = SettingModel.year()
-        user     = users.get_current_user()
-        selector = StudentModel.get_student_by_email(quarter, year, user.email())
-
-        if not selector:
-            return self.redirect('/partner')
-
-        # use selector info to find students in same lab section
-        selectees          = StudentModel.get_students_by_lab(quarter, year, selector.lab)
-        current_assignment = AssignmentModel.get_active_assign_with_latest_fade_in_date(quarter, year)
-
-        # if there are no assignments for this quarter, redirect to avoid errors
-        if not current_assignment:
-            return self.redirect('/partner?message=There are no assignments open for partner selection.')
-            
-        # get error message, if any
-        e = self.request.get('error')        
-
-        # check to see if partner selection period has closed
-        selection_closed = (datetime.now() - timedelta(hours=8) > current_assignment.close_date)
-
-        # get all current_partnerships for partnership status
-        partnerships = PartnershipModel.get_all_partnerships_for_assign(quarter, year, current_assignment.number)
-        partnerships = {p.initiator for p in partnerships} | {p.acceptor for p in partnerships}
-
-        # build dict to store information about partnership status
-        selectees = sorted({s.ucinetid: (s.key in partnerships,s) for s in selectees}.items(), key=lambda x: x[1][1].last_name)
-
-        # pass template values...
-        template_values = {
-            'error':            e,
-            'selector':         selector,                                  
-            'selectees':        selectees,                                
-            'selection_closed': selection_closed,
-            'current':          current_assignment,
-            'user':             user,
-            'sign_out':         users.create_logout_url('/'),
-        }
-        template = JINJA_ENV.get_template('/templates/partner_browse.html')
-        self.response.write(template.render(template_values))
 
 
 class ConfirmPartner(CustomHandler):
@@ -203,252 +126,6 @@ class ConfirmPartner(CustomHandler):
             return self.redirect('/admin/partners/add?message=' + message)
             
 
-
-class EditProfile(CustomHandler):
-
-    @login_required
-    def get(self):
-        template = JINJA_ENV.get_template('/templates/partner_update_profile.html')
-
-        quarter = SettingModel.quarter()
-        year    = SettingModel.year()
-        user    = users.get_current_user()
-        student = StudentModel.get_student_by_email(quarter, year, user.email())
-
-        if not student:
-            # redirect to main page if the student doesn't exist in the DB
-            return self.redirect('/partner')
-
-        programming_ability  = ['0: Never, or just a few times', '1: Occaisionally, but not regularly']
-        programming_ability += ['2: Regularly, but without much comfort or expertise']
-        programming_ability += ['3: Regularly, with comfortable proficiency', '4: Frequently and with some expertise']
-
-        template_values = {
-            'user':                user,
-            'sign_out':            users.create_logout_url('/'),
-            'student':             student,
-            'programming_ability': programming_ability,
-            'key':                 student.key.urlsafe(),
-        }
-        return self.response.write(template.render(template_values))
-
-
-    def post(self):
-        quarter = SettingModel.quarter()
-        year    = SettingModel.year()
-        user    = users.get_current_user()
-        student = StudentModel.get_student_by_email(quarter, year, user.email())
-
-        student.preferred_name      = self.request.get('preferred_name').strip()
-        bio                         = self.request.get('bio').strip()
-        student.bio                 = bio if bio != '' else student.bio         
-        availability                = self.request.get('availability').strip()
-        student.availability        = availability if availability != '' else student.availability 
-        programming_ability         = self.request.get('programming_ability')
-        student.programming_ability = programming_ability
-
-        # grab image and resize if necessary
-        avatar         = str(self.request.get('avatar'))
-        student.avatar = images.resize(avatar, 320, 320) if avatar != '' else student.avatar
-
-        phone_number         = self.request.get('phone_number')
-        student.phone_number = phone_number if phone_number != '000-000-0000' else student.phone_number
-        student.put()
-
-        # redirect to main page
-        return self.redirect('/partner/edit/profile')
-
-
-class EvaluatePartner(CustomHandler):
-    @login_required
-    def get(self):
-        quarter = SettingModel.quarter()
-        year    = SettingModel.year()
-
-        # get user
-        user = users.get_current_user()
-        # get student from user info
-        evaluator = StudentModel.get_student_by_email(quarter, year, user.email())
-        # get student's partner history
-        partners = PartnershipModel.get_active_partner_history_for_student(evaluator, quarter, year)
-        # grab the active eval assignments
-        eval_assigns = AssignmentModel.get_active_eval_assigns(quarter, year)
-        # grab partners for eval assignments
-        partners = [(eval_assign.number,PartnershipModel.get_partner_from_partner_history_by_assign(evaluator, partners, eval_assign.number)) for eval_assign in eval_assigns]
-        # filter out No Partner partnerships
-        partners = list(filter(lambda x: x[1] != "No Partner" and x[1] != None, partners))
-        eval_closed = len(eval_assigns) == 0
-    
-        rate20scale  = ["0 -- Never, completely inadequate", "1", "2", "3", "4"]
-        rate20scale += ["5 -- Seldom, poor quality", "6", "7", "8", "9"]
-        rate20scale += ["10 -- About half the time, average", "11", "12", "13", "14"]
-        rate20scale += ["15 -- Most of the time, good quality", "16", "17", "18", "19"]
-        rate20scale += ["20 -- Always, excellent"]
-
-        rate5scale  = ['1 - My partner was much more capable than I was']
-        rate5scale += ['2 - My partner was a little more capable than I was']
-        rate5scale += ['3 - We were about evenly matched, on average']
-        rate5scale += ['4 - I was a little more capable than my partner']
-        rate5scale += ['5 - I was a lot more capable than my partner']
-
-        rate10scale = [str(x / 10.0) for x in range(0, 105, 5)]
-
-        template_values = {
-            'eval_closed': eval_closed,
-            'rate_scale': rate20scale,
-            'rate5scale': rate5scale,
-            'rate10scale': rate10scale,
-            'partners': partners,
-        }
-        template = JINJA_ENV.get_template('/templates/partner_eval.html')
-        self.response.write(template.render(template_values))
-
-
-    def post(self):
-        quarter = SettingModel.quarter()
-        year    = SettingModel.year()
-        user    = users.get_current_user()
-
-        evaluator       = StudentModel.get_student_by_email(quarter, year, user.email())
-        partners        = PartnershipModel.get_active_partner_history_for_student(evaluator, quarter, year)
-        eval_key,num    = split_last(self.request.get('evaluatee'))
-        eval_assign     = int(num)
-        current_partner = ndb.Key(urlsafe=eval_key).get()
-
-        evaluations = EvalModel.get_existing_eval_by_assign(evaluator, current_partner, eval_assign)
-        for eval in evaluations:
-            eval.active = False
-            eval.put()
-
-        evaluation                   = Evaluation(evaluator = evaluator.key, evaluatee = current_partner.key)
-        evaluation.assignment_number = eval_assign
-        evaluation.year              = evaluator.year
-        evaluation.quarter           = evaluator.quarter
-        evaluation.active            = True
-        for i in range(1, 11):
-            evaluation.responses.append(self.request.get('q' + str(i)))
-        evaluation.put()
-
-        message  = 'Evaluation for ' + str(current_partner.last_name) + ', '
-        message += str(current_partner.first_name) + ' successfully submitted'
-
-        self.redirect('/partner?message=' + message)
-
-
-class HelpPage(CustomHandler):
-    def get(self):
-        user = users.get_current_user()
-        template_values = {}
-
-        if user:
-            template_values['user'] = user.email()
-            template_values['sign_out'] = users.create_logout_url('/')
-        else:
-            template_values['sign_in'] = users.create_login_url('/partner')
-
-        template = JINJA_ENV.get_template('/templates/partner_instructions.html')
-        self.response.write(template.render(template_values))
-
-
-class ImageHandler(CustomHandler):
-
-    def get(self, key):
-        # cast key from url from str to ndb.Key
-        key = ndb.Key(urlsafe=key)
-        # grab student associated w/ key and the corresponding avatar
-        image = key.get().avatar 
-        # set content type header...
-        self.response.headers['Content-Type'] = 'image/png'
-        # and 
-        return self.response.out.write(image)
-
-
-class MainPage(CustomHandler):
-    @login_required
-    def get(self):
-
-        # delcare page template
-        template = JINJA_ENV.get_template('/templates/partners_main.html')
-        # get current user
-        user = users.get_current_user()
-        student = None
-
-        try:
-            quarter = SettingModel.quarter()
-            year    = SettingModel.year()
-            # use user info to find student int DB
-            student = StudentModel.get_student_by_email(quarter, year, user.email())
-
-            self.session['quarter'] = quarter
-            self.session['year']    = year
-            self.session['student'] = student.key.urlsafe()
-
-            # get active assignments
-            active_assigns = sorted(AssignmentModel.get_active_assigns(quarter, year), key=lambda x: x.number)
-            # get active eval assigns
-            eval_assigns = sorted(AssignmentModel.get_active_eval_assigns(quarter, year), key=lambda x: x.number)
-            # find any active invitations for the current assignment that student has sent
-            sent_invitations = InvitationModel.get_sent_invites_by_student_and_mult_assigns(student, [x.number for x in active_assigns])
-            # find any active invitations for the current assignment that the student has received
-            received_invitations = InvitationModel.get_recvd_invites_by_student_and_mult_assigns(student, [x.number for x in active_assigns])
-            # find any partnerships in which the student has been involved
-            partners = PartnershipModel.get_active_partner_history_for_student(student, quarter, year)
-            partners = dict([(x.number,PartnershipModel.get_partner_from_partner_history_by_assign(student, partners, x.number)) for x in active_assigns])
-            # find evals the student has submitted
-            evals = EvalModel.get_eval_history_by_evaluator(student, True, quarter, year)
-            evals = dict([(x.assignment_number,x) for x in evals])
-            # get activity message, if any
-            message = self.request.get('message')
-            dropped = []
-            for x in active_assigns:
-                dropped += PartnershipModel.get_inactive_partnerships_by_student_and_assign(student, x.number).fetch()
-            dropped = sorted(dropped, key=lambda x: x.assignment_number)
-
-        except (AttributeError, IndexError):
-            template_values = {
-                'user': user,
-                'student': student,
-                'sign_out': users.create_logout_url('/'),
-                'email': user.email()
-            }
-            return self.response.write(template.render(template_values))
-
-        # pass template values...
-        template_values = {
-            'user': user,
-            'student': student,                             # ...student object
-            'active': active_assigns,
-            'evals': eval_assigns,
-            'submitted_evals': evals,
-            'sent_invitations': sorted(sent_invitations, key=lambda x: x.assignment_number),    # ...list of invitation objects
-            'received_invitations': sorted(received_invitations.items(), key=lambda x: x[0]),    # ...list of invitation objects
-            'partners': partners,                            # ...dict of partnership objects
-            'sign_out': users.create_logout_url('/'),        # sign out url
-            'message': message,
-            'profile': student.bio is None or student.availability is None or student.programming_ability is None,
-            'dropped': dropped,
-            'show_dropped': len(dropped) > 0 and len(filter(lambda x: x != None, partners.values())) < len(active_assigns),
-        }
-        self.response.write(template.render(template_values))
-
-
-
-class Main(webapp2.RequestHandler):
-    def get(self):
-        user = users.get_current_user()
-        template = JINJA_ENV.get_template('/templates/index.html')
-        template_values = {}
-
-        if user:
-            template_values['user'] = user.email()
-            template_values['sign_out'] = users.create_logout_url('/')
-        else:
-            template_values['sign_in'] = users.create_login_url('/partner')
-
-        self.response.write(template.render(template_values))
-
-
-
 class SelectPartner(CustomHandler):
     @login_required
     def get(self):
@@ -533,6 +210,3 @@ class SelectPartner(CustomHandler):
             return self.redirect('/partner?message=' + message)
 
 
-################################################################################
-################################################################################
-################################################################################
